split declaration files; better import
