--- test/types/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "esnext",
-		"module": "commonjs",
-		"lib": ["esnext", "dom"],
-		"noEmit": true,
-		"strict": true,
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
-		"noImplicitReturns": true,
-		"noFallthroughCasesInSwitch": true,
-		"allowSyntheticDefaultImports": true,
-		"esModuleInterop": true
-	}
-}
