+++ types/browser.d.ts
@@ -1,0 +1,9 @@
+import {NodeCGStaticBrowser} from './lib/nodecg-static';
+import {NodeCGBrowser} from './lib/nodecg-instance';
+
+declare const nodecg: NodeCGBrowser;
+declare const NodeCG: NodeCGStaticBrowser;
+
+export * from './lib/config';
+export * from './lib/logger';
+export * from './lib/replicant';
