+++ types/lib/platform.d.ts
@@ -1,0 +1,4 @@
+/**
+ * Used to select types for browser API or extensions API
+ */
+export type Platform = 'server' | 'browser';
