+++ types/server.d.ts
@@ -1,0 +1,7 @@
+export {NodeCGServer as NodeCG} from './lib/nodecg-instance';
+export {NodeCGStaticServer as NodeCGStatic} from './lib/nodecg-static';
+export {ReplicantServer as Replicant} from './lib/replicant';
+
+export * from './lib/config';
+export * from './lib/logger';
+export * from './lib/replicant';
